==> Note Books/7 Feature Engineering.ipynb <==
import pandas as pd
import numpy as np



data_path = "/home/danial/Data Science/Fraud Detection/Data/cleaned/creditcard_cleaned.csv"
df = pd.read_csv(data_path)

df.info()

# The 'Time' column is in seconds since the first transaction
# Convert it to hours and days for better patterns
df['Hour'] = (df['Time'] // 3600) % 24   # Hour of day
df['Day'] = df['Time'] // (3600 * 24)    # Day number since start

# Create a feature for weekday/weekend
df['Is_Weekend'] = df['Day'] % 7 >= 5

# --- Cell 2: Create ratio or derived features ---
# Log-transform the 'Amount' to reduce skewness
df['Log_Amount'] = np.log1p(df['Amount'])

# Example of normalization: amount per hour (relative spending pattern)
df['Amount_per_Hour'] = df['Amount'] / (df['Hour'] + 1)  # +1 to avoid division by zero


# --- Cell 3: Compute group-based statistics ---
# Calculate mean amount per day and map it back
day_mean_amount = df.groupby('Day')['Amount'].transform('mean')
df['Day_Amount_Diff'] = df['Amount'] - day_mean_amount


# --- Cell 4: Check new features ---
print(df[['Time','Hour','Day','Is_Weekend','Amount','Log_Amount','Amount_per_Hour','Day_Amount_Diff']].head())


# --- Cell 5: Save dataset with engineered features ---
df.to_csv('/home/danial/Data Science/Fraud Detection/Data/featured/featured.csv' , index=False)
